Quicksort, M O R E rectangles, faster shuffling, rectangle height changed to double, selection sort bigger delay
Hardcoded some delay to selection sort because it was too fast due to not ideal visualization and making user move the delay slider all the way to the right just to see selection sort a bit slower than the speed of light was a bad user experience. Changed rectangles heights from int to double to make them look better while their number grows

+++ SortingAlgorithms/Models/Rects.cs
@@ -8,4 +8,4 @@ public static class Rects
     public static ObservableCollection<Rectangle> Rectangles { get; } = new();
 }
 
-public record Rectangle (int Height, IBrush Color);
+public record Rectangle (double Height, IBrush Color);
